Split project integration tests from workspace model
