+++ docs/DECISIONS.md
@@ -211,6 +211,7 @@
 - Pure project loader coverage lives in focused test files, not `WorkspaceModelTests`. Direct tests for `ProjectInstructionLoader`, `LocalEnvironmentActionLoader`, `ProjectExtensionManifestLoader`, and `MemoryNoteLoader` should stay in their matching test files; `WorkspaceModelTests` should prove workspace integration only.
 - Workspace memory integration coverage lives in `WorkspaceMemoryIntegrationTests`, not `WorkspaceModelTests`. Global/project memory loading, `/remember`, agent memory tool execution, credential-like memory rejection, memory deletion, and the memory-add command are one feature group and should stay tested together outside the model monolith.
 - Workspace MCP integration coverage lives in `WorkspaceMCPIntegrationTests`, not `WorkspaceModelTests`. MCP lifecycle, dynamic tool catalog exposure, tool/resource/prompt execution from agent turns, and rejection of unadvertised MCP tools cross the workspace model, MCP runtime, tool cards, transcript events, and surfaces; keep those flows together outside the model monolith while lower-level MCP launcher/catalog/request tests stay in their focused files.
+- Workspace project integration coverage lives in `WorkspaceProjectIntegrationTests`, not `WorkspaceModelTests`. Project instruction snapshotting crosses the workspace model, metadata loader, thread context, agent submission, and surfaces; keep that flow outside the model monolith while direct instruction parsing stays in focused loader tests.
 - Runtime factory coverage lives in `WorkspaceRuntimeFactoryTests`, not `WorkspaceModelTests`. Tests for environment key detection, stored secret detection, deterministic mock override selection, and no-key model-catalog fallback should stay with the factory boundary; `WorkspaceModelTests` should cover runtime state only when the model applies or surfaces it.
 - Workspace project extension integration coverage lives in `WorkspaceProjectExtensionIntegrationTests`, not `WorkspaceModelTests`. Extension manifest loading into project/secondary-pane surfaces, update-command execution, metadata refresh, and success/failure transcript notices cross the workspace model and project metadata loader; keep those flows together outside the model monolith while low-level manifest parsing and shell-call shape tests stay focused.
 - Worktree request values, tool-call construction, and handoff transcript records live outside `WorkspaceModel`. `WorkspaceWorktreeRequests` owns the public create/remove request structs used by dialogs, SwiftUI, and desktop controllers. `WorkspaceWorktreeToolCallPlanner` owns canonical `host.git.worktree.create` and `host.git.worktree.remove` argument JSON, including branch/base trimming and force flags. `WorkspaceWorktreeOpenEngine` owns the pure local/SSH Remote `Worktree: ...` thread construction, notice payloads, and display labels. The workspace model owns only tool dispatch, local/SSH project registration, selection, persistence, and top-bar refresh after a worktree is created.
